Use not only single descriptors but the full (microdesc) archive.
This fixes the relay family calculation because relays are not
present in each relay descriptor document provided by CollecTor
(these only span a certain period of time).

Apart from the microdesc consensus, you have to provide a folder
structure as obtained by extracting one of the archives from
https://metrics.torproject.org/collector/archive/relay-descriptors/microdescs/
Note that the microdesc consensuses are included there, as well.